Move omnibar-help-center into interim-omnibar directory
This groups it with the other interim omnibar code so it can be
re-assessed as part of the switch to the permanent omnibar.
